uniform behaviour of option, command keys
So far, option, and command behaviour with the Apple internal keyboard
had been a little wonky (since karabiner config was geared towards my
external keyboards, where option and command keys are flipped). This
commit makes the behaviour uniform across both kinds of keyboards.

The change is commented out since I hardly ever use the built-in
keyboard, but it's great to know it works!